refactor: 리뷰 반영하여 수정

+++ src/main/java/calculator/Calculator.java
@@ -2,22 +2,22 @@ package calculator;
 
 public class Calculator {
 
-    int add(int a, int b) {
-        return a + b;
+    long add(int a, int b) {
+        return (long) a + b;
     }
 
-    int minus(int a, int b) {
-        return a - b;
+    long minus(int a, int b) {
+        return (long) a - b;
     }
 
-    int multiply(int a, int b) {
-        return a * b;
+    long multiply(int a, int b) {
+        return (long) a * b;
     }
 
-    int divide(int a, int b) {
+    long divide(int a, int b) {
         if (b == 0) {
             throw new IllegalArgumentException("0으로 나눌 수 없습니다.");
         }
-        return a / b;
+        return (long) a / b;
     }
 }
